requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.1
pydantic-core==2.33.0
sniffio==1.3.1
starlette==0.46.1
typing-extensions==4.13.0
typing-inspection==0.4.0
uvicorn==0.34.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
certifi==2025.1.31
httpcore==1.0.9
httpx==0.28.1
asyncpg==0.30.0
greenlet==3.1.1
sqlalchemy==2.0.40
Mako==1.3.9
MarkupSafe==3.0.2
alembic==1.15.2
aiosqlite==0.21.0
beanie==1.29.0
dnspython==2.7.0
lazy-model==0.2.0
motor==3.7.0
pymongo==4.12.1
toml==0.10.2